% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{abalone}
\format{
A data frame with 4177 rows and 9 variables:
\describe{
\item{\code{Sex}}{Male, Female, Infant}
\item{\code{Length}}{Longest shell measurement (mm)}
\item{\code{Diameter}}{Perpendicular to length (mm)}
\item{\code{Height}}{With meat in shell (mm)}
\item{\code{Whole_Weight}}{Whole abalone (g)}
\item{\code{Shuck_Weight}}{Weight of meat (g)}
\item{\code{Viscera_Weight}}{Gut weight after bleeding (g)}
\item{\code{Shell_Weight}}{Weight after being dried (g)}
\item{\code{Older_12}}{Whether the abalone is (strictly) greater than 12 years of age (0 = no, 1 = yes)}
}
}
\source{
Warwick J Nash, Tracy L Sellers, Simon R Talbot, Andrew J Cawthorn and Wes B Ford (1994) "The Population Biology of Abalone (\emph{Haliotis} species) in Tasmania. I. Blacklip Abalone (\emph{H. rubra}) from the North Coast and Islands of Bass Strait",Sea Fisheries Division, Technical Report No. 48 (ISSN 1034-3288)
}
\usage{
abalone
}
\description{
Predicting the age of abalone from physical measurements. The age of abalone is determined by cutting the shell through the cone, staining it, and counting the number of rings through a microscope -- a boring and time-consuming task. Other measurements, which are easier to obtain, are used to predict the age. Further information, such as weather patterns and location (hence food availability) may be required to solve the problem. Data comes from the UCI machine learning repository. Further information can be found at:  \url{https://archive.ics.uci.edu/ml/datasets/Abalone}
}
\keyword{datasets}
