% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{check_dataset}
\alias{check_dataset}
\title{Validate the input dataset before imputation}
\usage{
check_dataset(dataset)
}
\arguments{
\item{dataset}{The object passed to \code{misl()}.}
}
\description{
Validate the input dataset before imputation
}
\keyword{internal}
