% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{check_datatype}
\alias{check_datatype}
\title{Determine the outcome type of a column}
\usage{
check_datatype(x)
}
\arguments{
\item{x}{A vector (one column from the dataset).}
}
\value{
One of \code{"categorical"}, \code{"binomial"}, or \code{"continuous"}.
}
\description{
Determine the outcome type of a column
}
\keyword{internal}
