% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{Returns a complete tibble from the multiple imputed datasets}
\usage{
complete(misl_imputations, number)
}
\arguments{
\item{misl_imputations}{A list of misl imputed data objects}

\item{number}{Returns the \code{number} complete dataset.}
}
\value{
A tibble of the complete dataset
}
\description{
Returns a complete tibble from the multiple imputed datasets
}
\examples{

# This will generate imputations for the built-in nhanes dataset.
misl_imp <- misl(nhanes)

# You can then return the desired dataset
complete_dataset <- complete(misl_imp,1)

}
