% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{predict.misl_polr_fit}
\alias{predict.misl_polr_fit}
\title{Predict method for misl_polr_fit objects}
\usage{
\method{predict}{misl_polr_fit}(object, new_data, type = "prob", ...)
}
\description{
Predict method for misl_polr_fit objects
}
\keyword{internal}
