% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot the trace lines of the MISL algorithm}
\usage{
trace_plot(imputations)
}
\arguments{
\item{imputations}{Results of the MISL imputations}
}
\value{

}
\description{
Creates a traceplot using the imputations from MISL to view convergence. Of importance, these plots do not display categorical variables.
}
\examples{
# First generate the imputations
misl_imp <- misl(nhanes, m = 2, maxit = 2)

# Then you can generate the associated traceplot to investigate
trace_plot(misl_imp)
}
