% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Creates a traceplot using the imputations from MISL}
\usage{
traceplot(imputations)
}
\arguments{
\item{imputations}{Results of the MISL imputations}
}
\value{

}
\description{
Creates a traceplot using the imputations from MISL
}
