% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mode.R
\name{impute_mode}
\alias{impute_mode}
\title{Returns the mode (most occuring value)}
\usage{
impute_mode(x)
}
\arguments{
\item{x}{list}
}
\value{
num
}
\description{
Returns the mode (most occuring value)
}
\examples{
impute_mode(c(1,1,2,3))
}
