% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_placeholders.R
\name{impute_placeholders}
\alias{impute_placeholders}
\title{Imputes placeholders in the current dataset}
\usage{
impute_placeholders(dataset, column_number, missing_default)
}
\arguments{
\item{dataset}{The dataset to which placeholders are filled.}

\item{column_number}{The specific column within the dataset that is to be filled.}

\item{missing_default}{A string defining how placeholder values should be imputed with the misl algorithm. Allows for one of the following: \code{c("mean", "median")}. The default is \code{missing_default = "mean"}.}
}
\value{

}
\description{
Imputes placeholders in the current dataset
}
