% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_learners.R
\name{list_learners}
\alias{list_learners}
\title{Lists all learners available}
\usage{
list_learners(datatype)
}
\arguments{
\item{datatype}{A character vector of one of the following: "continuous", "binomial", "categorical"}
}
\value{
A list of learners compatible with the specified datatype
}
\description{
Lists all learners available for ensemble learning. This is an aliased function from SL3::sl3_list_learners
}
\examples{
list_learners("continuous")

}
