% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{misl}
\alias{misl}
\title{Imputes missing values using multiple imputation by super learning}
\usage{
misl(dataset)
}
\arguments{
\item{quiet}{Boolean}
}
\value{

}
\description{
Imputes missing values using multiple imputation by super learning
}
