% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{misl}
\alias{misl}
\title{Imputes missing values using multiple imputation by super learning}
\usage{
misl(
  dataset,
  m = 5,
  maxit = 5,
  seed = NA,
  con_method = c("Lrnr_mean", "Lrnr_glm"),
  bin_method = c("Lrnr_mean", "Lrnr_glm"),
  cat_method = c("Lrnr_mean", "Lrnr_glmnet"),
  missing_default = "mean",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Imputes missing values using multiple imputation by super learning
}
