% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{misl}
\alias{misl}
\title{Imputes missing values using multiple imputation by super learning}
\usage{
misl(
  dataset,
  m = 5,
  maxit = 5,
  seed = NA,
  con_method = c("Lrnr_mean", "Lrnr_glm_fast"),
  bin_method = c("Lrnr_mean", "Lrnr_glm_fast"),
  cat_method = c("Lrnr_mean"),
  ignore_predictors = NA,
  quiet = TRUE,
  delta_con = 0,
  delta_cat = 1
)
}
\arguments{
\item{dataset}{A dataframe or matrix containing the incomplete data. Missing values are represented with \code{NA}.}

\item{m}{The number of multiply imputed datasets to create. The default is \code{m=5}.}

\item{maxit}{The number of iterations for each of the \code{m} imputed datasets. The default is \code{maxit=5}.}

\item{seed}{Specify whether or not to include a seed for reproducible research. The default is \code{seed = NA}.}

\item{con_method}{A vector of strings to be supplied for building the super learner for columns containing continuous data. The default learners are \code{con_method = c("Lrnr_mean", "Lrnr_glm")}.}

\item{bin_method}{A vector of strings to be supplied for building the super learner for columns containing binomial data. Important to note that these values must only take on values \code{c(0,1,NA)}. The default learners are \code{bin_method = c("Lrnr_mean", "Lrnr_glm")}.}

\item{cat_method}{A vector of strings to be supplied for building the super learner for columns containing categorical data. The default learners are \code{bin_method = c("Lrnr_mean", "Lrnr_glmnet")}.}

\item{ignore_predictors}{A vector of strings to be supplied for ignoring in the prediction of other variables. The default is \code{ignore_predictors = NA}}

\item{quiet}{A boolean describing if progress of the misl algorithm should be printed to the console. The default is \code{quiet = TRUE}.}

\item{delta_con}{An integer to specify by how much continuous values should be shifted for the delta adjustmenet method of a sensitivity analysis. If the user does not specify a value, the imputations will not be augmented. The default is \code{delta_con = 0}.}

\item{delta_cat}{An integer to specify by how much binary/categorical values should be scaled for the delta adjustmenet method of a sensitivity analysis. If the user does not specify a value, the imputations will not be augmented. The default is \code{delta_cat = 1}.}
}
\value{
A list of \code{m} full tibbles.
}
\description{
Imputes missing values using multiple imputation by super learning
}
\examples{
# This will generate imputations for the built-in abalone dataset.
misl_imp <- misl(abalone,
m = 1,
abalone, maxit = 5, m = 5, quiet = TRUE,
con_method = c("Lrnr_glm_fast", "Lrnr_earth", "Lrnr_ranger"),
bin_method = c("Lrnr_earth", "Lrnr_glm_fast", "Lrnr_ranger"),
cat_method = c("Lrnr_independent_binomial", "Lrnr_ranger"))

}
