% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misl.R
\name{misl}
\alias{misl}
\title{Imputes missing values using multiple imputation by super learning}
\usage{
misl(
  dataset,
  m = 5,
  maxit = 5,
  seed = NA,
  con_method = c("Lrnr_mean", "Lrnr_glm_fast"),
  bin_method = c("Lrnr_mean", "Lrnr_glm_fast"),
  cat_method = c("Lrnr_mean"),
  ignore_predictors = NA,
  missing_default = "sample",
  quiet = TRUE
)
}
\arguments{
\item{dataset}{A dataframe or matrix containing the incomplete data. Missing values are represented with \code{NA}.}

\item{m}{The number of multiply imputed datasets to create. The default is \code{m=5}.}

\item{maxit}{The number of iterations for each of the \code{m} imputed datasets. The default is \code{maxit=5}.}

\item{seed}{Specify whether or not to include a seed for reproducible research. The default is \code{seed = NA}.}

\item{con_method}{A vector of strings to be supplied for building the super learner for columns containing continuous data. The default learners are \code{con_method = c("Lrnr_mean", "Lrnr_glm")}.}

\item{bin_method}{A vector of strings to be supplied for building the super learner for columns containing binomial data. Important to note that these values must only take on values \code{c(0,1,NA)}. The default learners are \code{bin_method = c("Lrnr_mean", "Lrnr_glm")}.}

\item{cat_method}{A vector of strings to be supplied for building the super learner for columns containing categorical data. The default learners are \code{bin_method = c("Lrnr_mean", "Lrnr_glmnet")}.}

\item{ignore_predictors}{A vector of strings to be supplied for ignoring in the prediction of other variables. The default is \code{ignore_predictors = NA}}

\item{missing_default}{A string defining how placeholder values should be imputed with the misl algorithm. Allows for one of the following: \code{c("mean", "median")}. The default is \code{missing_default = "mean"}.}

\item{quiet}{A boolean describing if progress of the misl algorithm should be printed to the console. The default is \code{quiet = TRUE}.}
}
\value{
A list of \code{m} full tibbles.
}
\description{
Imputes missing values using multiple imputation by super learning
}
\examples{
# This will generate imputations for the built-in nhanes dataset.
misl_imp <- misl(nhanes, m = 1)

}
