% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Creates a traceplot using the imputations from MISL}
\usage{
plot(imputations)
}
\arguments{
\item{imputations}{Results of the MISL imputations}
}
\value{

}
\description{
Creates a traceplot using the imputations from MISL
}
